refactor: restructure lambdas directory
Use esbuild to bundle lambdas. Use single package.json. Update deploy
scripts. Share code between lambdas.
