--- lambdas/auth-handler/.gitignore
@@ -1,3 +1,0 @@
-*.js
-utils/*.js
-*.zip
