--- lambdas/auth-handler/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "auth-handler",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.131.0",
-    "@aws-sdk/util-dynamodb": "^3.137.0",
-    "jwt-decode": "^3.1.2"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.101"
-  }
-}
