+++ lambdas/auth-handler/utils/check-for-user.ts
@@ -1,6 +1,6 @@
 import { GetItemCommand, GetItemCommandInput } from "@aws-sdk/client-dynamodb";
 import jwt_decode from "jwt-decode";
-import { ddbClient } from "../utils";
+import { ddbClient } from ".";
 
 export interface GoogleUserData {
   email: string;
