+++ lambdas/auth-handler/utils/create-user.ts
@@ -2,10 +2,10 @@ import { PutItemCommand } from "@aws-sdk/client-dynamodb";
 import { marshall } from "@aws-sdk/util-dynamodb";
 import { APIGatewayProxyEvent } from "aws-lambda";
 import jwt_decode from "jwt-decode";
-import { JsonResponse, buildResponse } from "./build-response";
 import { GoogleUserData } from "./check-for-user";
 import { ddbClient } from "./ddb-client";
 import { setAuthCookie } from "./set-auth-cookie";
+import { JsonResponse, buildResponse } from "../../utils";
 
 export const createUser = async (
   body: APIGatewayProxyEvent["body"],
