+++ lambdas/auth-handler/utils/index.ts
@@ -1,4 +1,3 @@
-export * from "./build-response";
 export * from "./check-for-user";
 export * from "./create-user";
 export * from "./is-user-authenticated";
