+++ lambdas/cors-handler/index.ts
@@ -3,10 +3,7 @@ import {
   APIGatewayProxyHandler,
   APIGatewayProxyResult,
 } from "aws-lambda";
-
-const DOMAIN_WHITELIST = ["https://gritlog.app", "https://stage.gritlog.app"];
-// eslint-disable-next-line security/detect-unsafe-regex
-const LOCALHOST_REGEX = /^http:\/\/localhost(:\d{1,5})?$/;
+import { isOriginAllowed } from "../utils";
 
 export const handler: APIGatewayProxyHandler = async (
   event: APIGatewayProxyEvent
@@ -29,14 +26,6 @@ export const handler: APIGatewayProxyHandler = async (
       message: "CORS validation passed: origin allowed.",
     }),
   };
-};
-
-const isOriginAllowed = (origin: string | undefined): boolean => {
-  if (origin === undefined) {
-    return false;
-  }
-
-  return DOMAIN_WHITELIST.includes(origin) || LOCALHOST_REGEX.test(origin);
 };
 
 const getHeaders = (origin: string) => ({
