--- lambdas/cors-handler/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "cors-handler",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/aws-lambda": "^8.10.114"
-  }
-}
