--- lambdas/db-handler/types.ts
@@ -1,10 +1,0 @@
-export interface Training {
-  readonly date: string;
-  readonly id: number;
-  readonly headline?: string | null | undefined;
-  readonly exercises: {
-    exerciseName?: string | null | undefined;
-    weight?: string | null | undefined;
-    repetitions?: string | null | undefined;
-  };
-}
