+++ lambdas/db-handler/utils/add-training.ts
@@ -2,10 +2,9 @@ import { PutItemCommand } from "@aws-sdk/client-dynamodb";
 import { marshall } from "@aws-sdk/util-dynamodb";
 import { APIGatewayProxyEvent } from "aws-lambda";
 import jwt_decode from "jwt-decode";
-import { JsonResponse, buildResponse } from "./build-response";
 import { GoogleUserData } from "./check-for-user";
 import { ddbClient } from "./ddb-client";
-import { Training } from "../types";
+import { JsonResponse, Training, buildResponse } from "../../utils";
 
 export const addTraining = async (
   jwt: string,
