+++ lambdas/db-handler/utils/index.ts
@@ -1,4 +1,3 @@
-export * from "./build-response";
 export * from "./check-for-user";
 export * from "./is-user-authenticated";
 export * from "./add-training";
