+++ scripts/db-handler-deploy-dev.sh
@@ -1,6 +1,6 @@
-tsc lambdas/db-handler/index.ts
-cd lambdas/db-handler && zip -r  db-handler.zip * -x *.ts utils/*.ts package-lock.json package.json   && cd ../..
-aws lambda update-function-code --function-name db-handler --region eu-central-1 --zip-file fileb://lambdas/db-handler/db-handler.zip --output text
+cd lambdas/ && npm run build &&
+cd db-handler/dist && zip -r db-handler.zip * && cd ../../.. &&
+aws lambda update-function-code --function-name db-handler --region eu-central-1 --zip-file fileb://lambdas/db-handler/dist/db-handler.zip --output text
 aws lambda wait function-updated --function-name db-handler
 aws lambda update-function-configuration --function-name db-handler  --output text
 aws lambda wait function-updated --function-name db-handler
