+++ tsconfig.json
@@ -21,6 +21,6 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true
   },
-  "include": ["src"],
+  "include": ["src", "lambdas/**/*.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
